Add files via upload
Cleaned up code and removed code that is not used and saved it in it's own file. Graph works for multiple stocks, instead of opening new windows for each stock. May leave it like this, since this will allow to graph multiple stocks and graph them in one chart to see how the may or may not relate.